+++ blockchain/genesis_test.go
@@ -34,7 +34,7 @@ func TestGenesis(t *testing.T) {
 
 	assert.Equal(uint32(1), genesisBlk.Header.version)
 	assert.Equal(uint32(1), genesisBlk.Header.chainID)
-	assert.Equal(uint32(0), genesisBlk.Header.height)
+	assert.Equal(uint64(0), genesisBlk.Header.height)
 	assert.Equal(uint64(1524676419), genesisBlk.Header.timestamp)
 	assert.Equal(expectedParentHash, genesisBlk.Header.prevBlockHash)
 	assert.Equal(uint32(1), genesisBlk.Header.trnxNumber)
